+++ RELEASE-NOTES.md
@@ -11,6 +11,7 @@
 * Adding `--image-name` option for image name argument in `spcs image-repository list-tags` for consistency with other commands.
 * Fixed errors during `spcs image-registry login` not being formatted correctly.
 * Project definition no longer accept extra fields. Any extra field will cause an error.
+* Changing imports in function/procedure section in `snowflake.yml` will cause the definition update on replace
 
 # v2.1.0